Windows bundles/settings

+++ init-bundles.vim
@@ -11,11 +11,15 @@ if has('unix')
 
     call plug#begin('~/.config/nvim/bundle')
 else
+    "if empty(glob("~\AppData\Local\nvim\autoload\plug.vim"))
+    "    silent call mkdir(expand("~\AppData\Local\nvim\autoload", 1), 'p')
+    "    execute '!curl -fLo ~\AppData\Local\nvim\autoload\plug.vim https://raw.github.com/junegunn/vim-plug/master/plug.vim'
+    "endif
+
     call plug#begin('~\AppData\Local\nvim\bundle')
 endif
 
 
-" original repos on github
 Plug 'Sirver/ultisnips'
 Plug 'christoomey/vim-tmux-navigator'
 Plug 'guns/xterm-color-table.vim', {'on': 'XtermColorTable'}
@@ -62,11 +66,11 @@ Plug 'tpope/vim-haml'
 Plug 'evanleck/vim-svelte'
 Plug 'zah/nim.vim', {'for': 'nim'}
 
-if !exists("$VSCODE")
+if !exists("$VSCODE") && !exists('g:vscode')
     Plug 'w0rp/ale'
 endif
 
-if !exists("$VSCODE") && !has("win32")
+if !exists("$VSCODE") && !exists("g:vscode") && !has("win32")
     Plug 'neoclide/coc.nvim', {'branch': 'release'}
     Plug 'marlonfan/coc-phpls', {'do': 'yarn install --frozen-lockfile'}
     Plug 'neoclide/coc-lists', {'do': 'yarn install --frozen-lockfile'}
